+++ MCA Portal Code/DirectorShareholdingsHiddenAttachment.py
@@ -9,95 +9,99 @@ import json
 import requests
 import mysql.connector
 from Config import create_main_config_dictionary
-
+import PyPDF2
+from PyPDF2 import PdfReader
 
 def update_value_in_db(db_config,name,no_of_shares,cin):
-    db_connection = mysql.connector.connect(**db_config)
-    db_cursor = db_connection.cursor()
-    db_connection.autocommit = True
-
-    check_name_query = "select * from authorized_signatories where cin = %s and LOWER(name) = %s and extracted_from = 'Master Data'"
-    values = (cin,str(name).lower())
-    print(check_name_query % values)
-    db_cursor.execute(check_name_query,values)
     try:
-        name_result = db_cursor.fetchone()
-    except Exception as e:
-        return
-    logging.info(name_result)
-    din = name_result[4]
-    designation = name_result[6]
-    shareholdings_query = "select * from director_shareholdings where cin = %s and LOWER(full_name) = %s"
-    values = (cin, str(name).lower())
-    print(shareholdings_query % values)
-    db_cursor.execute(shareholdings_query,values)
-    shareholdings_result = db_cursor.fetchall()
+        db_connection = mysql.connector.connect(**db_config)
+        db_cursor = db_connection.cursor()
+        db_connection.autocommit = True
 
-    paid_up_capital_query = "select paidup_capital from shareholdings_summary where cin = %s"
-    paid_up_capital_value = (cin,)
-    logging.info(paid_up_capital_query % paid_up_capital_value)
-    db_cursor.execute(paid_up_capital_query,paid_up_capital_value)
-    paid_up_capital = db_cursor.fetchone()[0]
-    try:
-        percentage_holding = (float(no_of_shares)/float(paid_up_capital))*100
-        percentage_holding = round(float(percentage_holding),2)
-        logging.info(percentage_holding)
-    except Exception as e:
-        logging.info(f"Error in calculating percentage holding {e}")
-        percentage_holding = None
-    try:
-        year_query = "select * from director_shareholdings where cin = %s and din_pan = ''"
-        year_values = (cin,)
-        logging.info(year_query % year_values)
-        db_cursor.execute(year_query,year_values)
-        year_result = db_cursor.fetchone()
-        year = year_result[6]
-        financial_year = year_result[7]
-    except Exception as e:
-        logging.info(f"Error in capturing year and financial year {e}")
-        year = ''
-        financial_year = ''
+        check_name_query = "select * from authorized_signatories where cin = %s and LOWER(name) = %s and extracted_from = 'Master Data'"
+        values = (cin,str(name).lower())
+        print(check_name_query % values)
+        db_cursor.execute(check_name_query,values)
+        try:
+            name_result = db_cursor.fetchone()
+        except Exception as e:
+            return
+        logging.info(name_result)
+        din = name_result[4]
+        designation = name_result[6]
+        shareholdings_query = "select * from director_shareholdings where cin = %s and LOWER(full_name) = %s"
+        values = (cin, str(name).lower())
+        print(shareholdings_query % values)
+        db_cursor.execute(shareholdings_query,values)
+        shareholdings_result = db_cursor.fetchall()
 
-    if len(name_result) != 0:
-        if len(shareholdings_result) != 0:
-            update_query = "UPDATE director_shareholdings set no_of_shares = %s where cin = %s and LOWER(full_name) = %s"
-            update_values = (no_of_shares,cin,str(name).lower())
-            logging.info(update_query % update_values)
-            db_cursor.execute(update_query,update_values)
-
-            din_update_query = "UPDATE director_shareholdings set din_pan = %s where cin = %s and LOWER(full_name) = %s"
-            din_update_values = (din, cin, str(name).lower())
-            logging.info(din_update_query % din_update_values)
-            db_cursor.execute(din_update_query, din_update_values)
-
-            percentage_holding_query = "UPDATE director_shareholdings set percentage_holding = %s where cin = %s and LOWER(full_name) = %s"
-            percentage_holding_value = (percentage_holding, cin, str(name).lower())
-            logging.info(percentage_holding_query % percentage_holding_value)
-            db_cursor.execute(percentage_holding_query,percentage_holding_value)
-
-            designation_query = "UPDATE director_shareholdings set designation = %s where cin = %s and LOWER(full_name) = %s"
-            designation_value = (designation, cin, str(name).lower())
-            logging.info(designation_query % designation_value)
-            db_cursor.execute(designation_query, designation_value)
-
-            year_query = "UPDATE director_shareholdings set year = %s where cin = %s and LOWER(full_name) = %s"
-            year_values = (year,cin,str(name).lower())
+        paid_up_capital_query = "select paidup_capital from shareholdings_summary where cin = %s"
+        paid_up_capital_value = (cin,)
+        logging.info(paid_up_capital_query % paid_up_capital_value)
+        db_cursor.execute(paid_up_capital_query,paid_up_capital_value)
+        paid_up_capital = db_cursor.fetchone()[0]
+        try:
+            percentage_holding = (float(no_of_shares)/float(paid_up_capital))*100
+            percentage_holding = round(float(percentage_holding),2)
+            logging.info(percentage_holding)
+        except Exception as e:
+            logging.info(f"Error in calculating percentage holding {e}")
+            percentage_holding = None
+        try:
+            year_query = "select * from director_shareholdings where cin = %s and din_pan = ''"
+            year_values = (cin,)
             logging.info(year_query % year_values)
             db_cursor.execute(year_query,year_values)
+            year_result = db_cursor.fetchone()
+            year = year_result[6]
+            financial_year = year_result[7]
+        except Exception as e:
+            logging.info(f"Error in capturing year and financial year {e}")
+            year = ''
+            financial_year = ''
 
-            financial_year_query = "UPDATE director_shareholdings set financial_year = %s where cin = %s and LOWER(full_name) = %s"
-            financial_year_values = (financial_year, cin, str(name).lower())
-            logging.info(financial_year_query % financial_year_values)
-            db_cursor.execute(financial_year_query, financial_year_values)
+        if len(name_result) != 0:
+            if len(shareholdings_result) != 0:
+                update_query = "UPDATE director_shareholdings set no_of_shares = %s where cin = %s and LOWER(full_name) = %s"
+                update_values = (no_of_shares,cin,str(name).lower())
+                logging.info(update_query % update_values)
+                db_cursor.execute(update_query,update_values)
 
-        else:
-            insert_query = "INSERT INTO director_shareholdings(cin,full_name,no_of_shares,din_pan,percentage_holding,designation,year,financial_year) VALUES (%s,%s,%s,%s,%s,%s,%s,%s)"
-            insert_values = (cin,name,no_of_shares,din,percentage_holding,designation,year,financial_year)
-            print(insert_query % insert_values)
-            db_cursor.execute(insert_query,insert_values)
+                din_update_query = "UPDATE director_shareholdings set din_pan = %s where cin = %s and LOWER(full_name) = %s"
+                din_update_values = (din, cin, str(name).lower())
+                logging.info(din_update_query % din_update_values)
+                db_cursor.execute(din_update_query, din_update_values)
 
-    db_cursor.close()
-    db_connection.close()
+                percentage_holding_query = "UPDATE director_shareholdings set percentage_holding = %s where cin = %s and LOWER(full_name) = %s"
+                percentage_holding_value = (percentage_holding, cin, str(name).lower())
+                logging.info(percentage_holding_query % percentage_holding_value)
+                db_cursor.execute(percentage_holding_query,percentage_holding_value)
+
+                designation_query = "UPDATE director_shareholdings set designation = %s where cin = %s and LOWER(full_name) = %s"
+                designation_value = (designation, cin, str(name).lower())
+                logging.info(designation_query % designation_value)
+                db_cursor.execute(designation_query, designation_value)
+
+                year_query = "UPDATE director_shareholdings set year = %s where cin = %s and LOWER(full_name) = %s"
+                year_values = (year,cin,str(name).lower())
+                logging.info(year_query % year_values)
+                db_cursor.execute(year_query,year_values)
+
+                financial_year_query = "UPDATE director_shareholdings set financial_year = %s where cin = %s and LOWER(full_name) = %s"
+                financial_year_values = (financial_year, cin, str(name).lower())
+                logging.info(financial_year_query % financial_year_values)
+                db_cursor.execute(financial_year_query, financial_year_values)
+
+            else:
+                insert_query = "INSERT INTO director_shareholdings(cin,full_name,no_of_shares,din_pan,percentage_holding,designation,year,financial_year) VALUES (%s,%s,%s,%s,%s,%s,%s,%s)"
+                insert_values = (cin,name,no_of_shares,din,percentage_holding,designation,year,financial_year)
+                print(insert_query % insert_values)
+                db_cursor.execute(insert_query,insert_values)
+
+        db_cursor.close()
+        db_connection.close()
+    except Exception as e:
+        logging.info(f"Exception {e} occured while inserting into db")
 
 
 def image_to_text(image_path):
@@ -114,7 +118,7 @@ def get_hidden_attachment(input_pdf_path, output_path,file_name_hidden_pdf):
         item_name_dict[each_item] = doc.embfile_info(each_item)["filename"]
 
     for item_name, file_name in item_name_dict.items():
-        if 'shareholders' in str(file_name).lower() or 'shareholder' in str(file_name).lower():
+        if 'shareholders' in str(file_name).lower() or 'shareholder' in str(file_name).lower() or 'share holders' in str(file_name).lower():
             out_pdf =  output_path + "\\" + file_name
             logging.info(out_pdf)
             fData = doc.embfile_get(item_name)
@@ -172,22 +176,28 @@ def MGT_director_shareholdings_pdf_to_db(pdf_path,config_dict,db_config,cin):
             xml_file.write("<?xml version='1.0' encoding='utf-8'?>\n")
             xml_file.write("<PDFData>\n")
         total_text = ''
-        for page_num in range(pdf_document.page_count):
-            page = pdf_document.load_page(page_num)
-            for img_index, image in enumerate(page.get_images(full=True)):
-                try:
-                    xref = image[0]
-                    base_image = pdf_document.extract_image(xref)
-                    image_data = base_image["image"]
+        text = ''
+        pdf_reader = PdfReader(pdf_path)
+        for page in pdf_reader.pages:
+            text += page.extract_text()
+        total_text = text
+        if total_text == '':
+            for page_num in range(pdf_document.page_count):
+                page = pdf_document.load_page(page_num)
+                for img_index, image in enumerate(page.get_images(full=True)):
+                    try:
+                        xref = image[0]
+                        base_image = pdf_document.extract_image(xref)
+                        image_data = base_image["image"]
 
-                    with open(f"temp_image_{img_index}.png", "wb") as img_file:
-                        img_file.write(image_data)
-                    text = image_to_text(f"temp_image_{img_index}.png")
-                    total_text += text
-                except Exception as e:
-                    print(f"Exception Occured while converting image to text{e}")
-                else:
-                    os.remove(f"temp_image_{img_index}.png")
+                        with open(f"temp_image_{img_index}.png", "wb") as img_file:
+                            img_file.write(image_data)
+                        text = image_to_text(f"temp_image_{img_index}.png")
+                        total_text += text
+                    except Exception as e:
+                        print(f"Exception Occured while converting image to text{e}")
+                    else:
+                        os.remove(f"temp_image_{img_index}.png")
         print(total_text)
         shareholders_details = fetch_address_din_using_open_ai(total_text,config_dict)
         print(shareholders_details)
@@ -220,7 +230,7 @@ def mgt_director_shareholdings_main(db_config,config_dict,output_directory,pdf_p
     else:
         return True
 
-# pdf_path = r"C:\Users\BRADSOL123\Desktop\XBRL\Form MGT-7A-22102023_signed.pdf"
+# pdf_path = r"C:\Users\BRADSOL123\Desktop\XBRL\SD brakes\Form MGT-7A-31102022.pdf"
 # excel_path = r"C:\Users\BRADSOL123\Documents\Python\Config\Config_Python.xlsx"
 # sheet_name = 'OpenAI'
 # config_dict,status = create_main_config_dictionary(excel_path,sheet_name)
@@ -230,6 +240,6 @@ def mgt_director_shareholdings_main(db_config,config_dict,output_directory,pdf_p
 # "password": "o2i=hi,64u*I",
 # "database": "classle3_mns_credit",
 # }
-# cin = 'U25112MP1992PTC007003'
-# output_directory = r"C:\Users\BRADSOL123\Desktop\XBRL"
+# cin = 'U35203HR2011PTC043528'
+# output_directory = r"C:\Users\BRADSOL123\Desktop\XBRL\SD brakes"
 # mgt_director_shareholdings_main(db_config,config_dict,output_directory,pdf_path,cin)
